+++ resources/aws_rds_global_cluster.rb
@@ -1,7 +1,6 @@
 # Import API specifics
 use "awscc_base"
 
-unified_mode true
 resource_name :aws_rds_global_cluster
 provides :aws_rds_global_cluster, target_mode: true, platform: "aws"