+++ resources/aws_route53_cidr_collection.rb
@@ -1,7 +1,6 @@
 # Import API specifics
 use "awscc_base"
 
-unified_mode true
 resource_name :aws_route53_cidr_collection
 provides :aws_route53_cidr_collection, target_mode: true, platform: "aws"
 
@@ -23,6 +22,7 @@ property :locations, Array,
 
 property :name, String,
          name_property: true,
+         required: true,
          callbacks: {
            "name is not a String" => lambda { |v| v.is_a? String },
            "name needs to be 1..64 characters" => lambda { |v| v.length >= 1 && v.length <= 64 },